Rename JS modules to .mjs
