+++ aidants_connect/importmaps.py
@@ -3,6 +3,7 @@ from django.conf import settings
 from importmap import static
 
 importmaps = {
-    "AidantsConnectApplication": static("js/ac-app.mjs"),
     "Stimulus": settings.STIMULUS_JS_URL.replace("umd.js", "js"),
+    "AidantsConnectApplication": static("js/ac-app.mjs"),
+    "MarkdownEditor": static("js/widgets/markdown-editor.mjs"),
 }
