+++ aidants_connect_common/static/js/formation-informations.mjs
