+++ aidants_connect_common/templates/formation/formation-registration.html
@@ -142,5 +142,5 @@
 {% block extrajs %}
   {% stimulusjs %}
   <script defer src="{% url 'js_reverse' %}"></script>
-  <script defer type="module" src="{% static 'js/formation-informations.js' %}"></script>
+  <script defer type="module" src="{% static 'js/formation-informations.mjs' %}"></script>
 {% endblock extrajs %}
