+++ aidants_connect_common/templates/habilitation/follow-my-request.html
@@ -51,5 +51,5 @@
   </template>
   {% stimulusjs %}
   <script defer src="{% url 'js_reverse' %}"></script>
-  <script defer type="module" src="{% static 'js/follow-request-modale.js' %}"></script>
+  <script defer type="module" src="{% static 'js/follow-request-modale.mjs' %}"></script>
 </div>
