+++ aidants_connect_habilitation/templates/add_aidants_request.html
@@ -1,9 +1,9 @@
 {% extends 'layouts/main-habilitation.html' %}
-{% load static form_extras ac_common %}
+{% load ac_common form_extras static %}
 
 {% block title %}
   Aidants Connect — Demande d'habilitation pour {{ organisation.name }}
-{% endblock %}
+{% endblock title %}
 
 {% block content %}
   <div class="fr-container" data-controller="personnel-form">
@@ -60,9 +60,9 @@
       {% include "_aidant_form.html" with form=form.empty_form %}
     </template>
   </div>
-{% endblock %}
+{% endblock content %}
 
 {% block extrajs %}
   {% stimulusjs %}
-  <script defer type="module" src="{% static 'js/personnel-form.js' %}"></script>
-{% endblock %}
+  <script defer type="module" src="{% static 'js/personnel-form.mjs' %}"></script>
+{% endblock extrajs %}
