+++ aidants_connect_habilitation/templates/organisation_form.html
@@ -1,9 +1,9 @@
 {% extends 'layouts/main-habilitation.html' %}
-{% load static form_extras ac_common %}
+{% load ac_common form_extras static %}
 
 {% block title %}
   Aidants Connect - Structure
-{% endblock %}
+{% endblock title %}
 
 {% block content %}
   <div class="fr-container">
@@ -102,13 +102,13 @@
     </div>
     {% include "_more-info.html" %}
   </div>
-{% endblock %}
+{% endblock content %}
 
 {% block extrajs %}
   {% stimulusjs %}
-  <script defer type="module" src="{% static 'js/organisation-form.js' %}"></script>
-{% endblock %}
+  <script defer type="module" src="{% static 'js/organisation-form.mjs' %}"></script>
+{% endblock extrajs %}
 
 {% block extracss %}
   <link href="{% static 'css/personnel_form.css' %}" rel="stylesheet">
-{% endblock %}
+{% endblock extracss %}
