+++ aidants_connect_pico_cms/static/js/widgets/markdown-editor-init.mjs
@@ -1,0 +1,8 @@
+import {aidantsConnectApplicationReady} from "AidantsConnectApplication"
+import {MarkdownEditor} from "MarkdownEditor"
+
+aidantsConnectApplicationReady.then(application => {
+    const container = document.querySelector("[data-markdown-editor-target='textareaContainer']").parentNode;
+    container.dataset.controller = "markdown-editor";
+    application.register("markdown-editor", MarkdownEditor);
+});
