+++ package.json
@@ -2,7 +2,8 @@
   "devDependencies": {
     "sass": "^1.79.4",
     "Stimulus": "npm:stimulus@^3.2.2",
-    "AidantsConnectApplication": "file:./aidants_connect_common/static/js/ac-app.mjs"
+    "AidantsConnectApplication": "file:./aidants_connect_common/static/js/ac-app.mjs",
+    "MarkdownEditor": "file:./aidants_connect_pico_cms/static/js/widgets/markdown-editor.js"
   },
   "imports": {
     "Stimulus": "@hotwired/stimulus"
